Made lots of changes.  Latest change was to make Checkout Activity neater using Jetpack Compose.

+++ app/src/main/java/com/tritongames/shoppingwishlist/data/models/bestbuy/Product.kt
@@ -285,7 +285,7 @@ data class Product(
     @SerializedName("shortDescription")
     val shortDescription: Any,
     @SerializedName("sku")
-    val sku: Int,
+    val sku: Long,
     @SerializedName("softwareAge")
     val softwareAge: Any,
     @SerializedName("softwareGrade")
